# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AlasGlobalTour API
  version: 1.0.0
  description: >
    API oficial de la plataforma ALAS Global Tour. Gestiona circuitos, eventos,
    competidores, inscripciones, pagos, rankings y contenido editorial.
    Todos los endpoints que expongan puntajes de SurfScores deben incluir
    la atribución "Results by SurfScores.com" en la respuesta.
  contact:
    email: dev@alasglobaltour.com
  license:
    name: Proprietary
servers:
  - url: https://api.alasglobaltour.com/v1
    description: Producción
  - url: https://api-staging.alasglobaltour.com/v1
    description: Staging
components:
  securitySchemes:
    bearerAuth: # Este es un nombre arbitrario
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Introduce el token JWT de Firebase aquí."
security:
  - bearerAuth: []
...
paths:
...
  /events/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: Obtener evento por ID
      responses:
        '200':
          description: Evento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Events]
      summary: Actualizar evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '200':
          description: Evento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Events]
      summary: Eliminar evento
      responses:
        '204':
          description: Evento eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No se puede eliminar un evento con inscripciones confirmadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /events/{eventId}/categories:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: Categorías habilitadas para el evento (con cupo y tarifas efectivas)
      responses:
        '200':
          description: Categorías del evento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCategoryListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Events]
      summary: Actualizar categorías y tarifas habilitadas para el evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categories]
              properties:
                useCircuitTariffs:
                  type: boolean
                  description: true hereda tarifas del circuito; false usa tarifas personalizadas por categoría
                categories:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventCategoryRequest'
      responses:
        '200':
          description: Categorías actualizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCategoryListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /events/{eventId}/results:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: Resultados del evento por categoría
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Resultados con atribución obligatoria de SurfScores
          content:
            application/json:
              schema:
                type: object
                properties:
                  attribution:
                    type: string
                    example: Results by SurfScores.com
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventResultResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Events]
      summary: Registrar resultados de un evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, results]
              properties:
                categoryId:
                  type: string
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventResultRequest'
      responses:
        '201':
          description: Resultados registrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /events/{eventId}/prize-distribution:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: Distribución de premios según el nivel de estrellas del evento
      responses:
        '200':
          description: Tabla de premios
          content:
            application/json:
              schema:
                type: object
                properties:
                  stars:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrizeDistributionRow'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ═══════════════════════════ CATEGORIES ═══════════════════════════
...
  /admin/users:
    get:
      tags: [Admin Users]
      summary: Listar usuarios administrativos
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/AdminRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AdminUserStatus'
      responses:
        '200':
          description: Lista de usuarios administrativos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Admin Users]
      summary: Crear usuario administrativo con contraseña temporal y correo de invitación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: El correo ya está en uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Admin Users]
      summary: Obtener usuario administrativo
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Admin Users]
      summary: Actualizar rol o estado del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Admin Users]
      summary: Eliminar usuario administrativo
      responses:
        '204':
          description: Usuario eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No se puede eliminar el propio usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── PARAMETERS ───────────────────────────────────────────────
  parameters:
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        example: evt-rbc-2026
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        example: usr-001

  # ─── REUSABLE RESPONSES ───────────────────────────────────────
  responses:
    BadRequest:
      description: Error de validación en la solicitud
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ─── SCHEMAS ──────────────────────────────────────────────────
  schemas:
...
    EventStatusPublic:
      type: string
      enum: [Inscripciones Abiertas, Próximamente, Completado, Cerrado]
    EventStatusAdmin:
      type: string
      enum: [Activo, Próximamente, Completado, Cancelado, Borrador]
    EventAccessType:
        type: string
        enum: [Abierto, Restringido, Solo invitación]
    EventType:
        type: string
        enum: [Regular, Prime, SuperPrime]
...
    CategoryGender:
      type: string
      enum: [Masculino, Femenino, Ambos]
...
    AdminRole:
      type: string
      enum: [Super Admin, Admin, Árbitro, Revisor]
    AdminUserStatus:
      type: string
      enum: [Activo, Inactivo]
...
    EventRequest:
      type: object
      required: [nombre, circuitId, fechaInicio, fechaFin, pais, ciudad, playa, stars]
      properties:
        nombre:
          type: string
          example: Roca Bruja Classic
        circuitId:
          type: string
          example: open26
        fechaInicio:
          type: string
          format: date
          example: '2026-07-12'
        fechaFin:
          type: string
          format: date
          example: '2026-07-15'
        pais:
          type: string
          example: Perú
        ciudad:
          type: string
          example: Chicama
        playa:
          type: string
          example: Roca Bruja
        auspiciador:
          type: string
          nullable: true
          example: Monster Energy
        stars:
          type: integer
          minimum: 1
          maximum: 7
          example: 3
        capacidadMaxima:
          type: integer
          example: 60
        prizeAmountUsd:
          type: number
          format: float
          example: 5000
        imagenUrl:
          type: string
          nullable: true
          example: https://cdn.alas.app/events/roca-bruja-2026-poster.jpg
        surfScoresCode:
          type: string
          example: RBC-2026-PER
        eventType:
          $ref: '#/components/schemas/EventType'
          default: Regular
        accessType:
          $ref: '#/components/schemas/EventAccessType'
        estado:
          $ref: '#/components/schemas/EventStatusAdmin'
          default: Borrador
    EventResponse:
      allOf:
        - $ref: '#/components/schemas/EventRequest'
        - type: object
          properties:
            id:
              type: string
              example: evt-rbc-2026
            enrolledCount:
              type: integer
            statusPublic:
              $ref: '#/components/schemas/EventStatusPublic'
            lugar:
              type: string
              example: Chicama, Perú
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    EventCategoryRequest:
      type: object
      required: [categoryId]
      properties:
        categoryId:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 7
          nullable: true
          description: null hereda el nivel de estrellas del evento
        customTariffUsd:
          type: number
          format: float
          nullable: true
          description: null hereda tarifa del circuito
        capacidad:
          type: integer
          nullable: true
    EventCategoryResponse:
      allOf:
        - $ref: '#/components/schemas/EventCategoryRequest'
        - type: object
          properties:
            categoryName:
              type: string
            gender:
              $ref: '#/components/schemas/CategoryGender'
            effectiveTariffUsd:
              type: number
              format: float
            enrolledCount:
              type: integer
    EventCategoryListResponse:
      type: object
      properties:
        useCircuitTariffs:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/EventCategoryResponse'

    # ── CATEGORIES ────────────────────────────────────────────────
...
    EventResultRequest:
      type: object
      required: [competitorId, place, ligaPoints]
      properties:
        competitorId:
          type: string
        place:
          type: string
          example: '1°'
        ligaPoints:
          type: integer
        prizeUsd:
          type: number
          format: float
          nullable: true
        heatOla1:
          type: number
          format: float
          nullable: true
        heatOla2:
          type: number
          format: float
          nullable: true
    EventResultResponse:
      allOf:
        - $ref: '#/components/schemas/EventResultRequest'
        - type: object
          properties:
            id:
              type: string
            competitorName:
              type: string
            country:
              type: string
            heatScoreTotal:
              type: number
              format: float
              nullable: true
              description: Suma de ola1 + ola2
    PrizeDistributionRow:
      type: object
      properties:
        placeLabel:
          type: string
          example: '1°'
        prizeUsd:
          type: number
          format: float
          description: Premio en USD para el nivel de estrellas del evento

    # ── ARTICLES ──────────────────────────────────────────────────
...
    AdminUserRequest:
      type: object
      required: [nombre, apellido, email, rol]
      properties:
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/AdminRole'
        sendInvitationEmail:
          type: boolean
          default: true
    AdminUserUpdateRequest:
      type: object
      properties:
        rol:
          $ref: '#/components/schemas/AdminRole'
        status:
          $ref: '#/components/schemas/AdminUserStatus'
    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
        initials:
          type: string
          example: GV
        fullName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/AdminRole'
        status:
          $ref: '#/components/schemas/AdminUserStatus'
        lastSession:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    AdminUserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserResponse'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time